Exceptions in cart_item_repository.py and cart_item_service.py done.

+++ app/offices/repositories/office_repository.py
@@ -1,9 +1,8 @@
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 from app.offices.models import Office
+from app.offices.exceptions import *
 
-
-# from app.offices.exceptions import *
 
 class OfficeRepository:
 
@@ -19,11 +18,13 @@ class OfficeRepository:
             return office
         except IntegrityError as e:
             raise e
+        except Exception as e:
+            raise e
 
     def read_by_id(self, office_id: str):
         office = self.db.query(Office).filter(Office.office_id == office_id).first()
-        # if office is None:
-        #     raise EmployeeNotFoundException(f"Employee with provided ID: {employee_id} not found.", 400)
+        if office is None:
+            raise OfficeNotFoundException(f"Office with provided id - {office_id} not found.", 400)
         return office
 
     def read_all(self):
@@ -33,8 +34,8 @@ class OfficeRepository:
     def delete_by_id(self, office_id: str):
         try:
             office = self.db.query(Office).filter(Office.office_id == office_id).first()
-            # if office is None:
-            #     raise EmployeeNotFoundException(f"Employee with provided ID: {employee_id} not found.", 400)
+            if office is None:
+                raise OfficeNotFoundException(f"Office with provided id - {office_id} not found.", 400)
             self.db.delete(office)
             self.db.commit()
             return True
@@ -45,8 +46,8 @@ class OfficeRepository:
                country: str = None, postal_code: str = None, territory: str = None):
         try:
             office = self.db.query(Office).filter(Office.office_id == office_id).first()
-            # if office is None:
-            #     raise EmployeeNotFoundException(f"Employee with provided ID: {employee_id} not found.", 400)
+            if office is None:
+                raise OfficeNotFoundException(f"Office with provided id - {office_id} not found.", 400)
             if name is not None:
                 office.name = name
             if phone is not None:
@@ -67,35 +68,3 @@ class OfficeRepository:
             return office
         except Exception as e:
             raise e
-
-
-    # def update_name(self, product_category_id: str, new_name: str):
-    #     try:
-    #         product_category = self.db.query(ProductCategory).filter(ProductCategory.product_category_id ==
-    #                                                                  product_category_id).first()
-    #         product_category.name = new_name
-    #         self.db.add(product_category)
-    #         self.db.commit()
-    #         self.db.refresh(product_category)
-    #         return product_category
-    #     except Exception as e:
-    #         raise e
-    #
-    # def read_by_name(self, name: str):
-    #     product_category = self.db.query(ProductCategory).filter(ProductCategory.name == name).first()
-    #     return product_category
-    #
-    # # -----------------------------------------------------------------------
-    #
-    # def get_employees_by_characters(self, characters: str):
-    #     employees = self.db.query(Employee).filter(Employee.name.like(characters + "%")).all()
-    #     if employees is None:
-    #         raise EmployeeNotFoundException(f"Employee with provided characters: {characters} not found.", 400)
-    #     return employees
-    #
-    # def get_employees_by_employee_type_id(self, employee_type_id: str):
-    #     employees = self.db.query(Employee).filter(Employee.employee_type_id == employee_type_id).first()
-    #     if employees is None:
-    #         raise EmployeeNotFoundException(f"Employee with provided employee type id: {employee_type_id} not found.",
-    #                                         400)
-    #     return employees
